Move REST model tests
