+++ backend/plugins/rest/src/main/java/org/optaweb/vehiclerouting/plugin/rest/model/PortableLocation.java
@@ -38,7 +38,7 @@ public class PortableLocation {
 
     private final String description;
 
-    static PortableLocation fromLocation(Location location) {
+    public static PortableLocation fromLocation(Location location) {
         Objects.requireNonNull(location, "location must not be null");
         return new PortableLocation(
                 location.id(),
@@ -48,7 +48,7 @@ public class PortableLocation {
     }
 
     @JsonCreator
-    PortableLocation(
+    public PortableLocation(
             @JsonProperty(value = "id") long id,
             @JsonProperty(value = "lat") BigDecimal latitude,
             @JsonProperty(value = "lng") BigDecimal longitude,
