+++ backend/plugins/rest/src/test/resources/org/optaweb/vehiclerouting/plugin/rest/model/portable-location.json
