--- backend/plugins/websocket/.gitignore
@@ -1,1 +1,0 @@
-/target
