--- backend/plugins/websocket/pom.xml
@@ -1,63 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-  ~ Copyright 2021 Red Hat, Inc. and/or its affiliates.
-  ~
-  ~ Licensed under the Apache License, Version 2.0 (the "License");
-  ~ you may not use this file except in compliance with the License.
-  ~ You may obtain a copy of the License at
-  ~
-  ~     http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing, software
-  ~ distributed under the License is distributed on an "AS IS" BASIS,
-  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~ See the License for the specific language governing permissions and
-  ~ limitations under the License.
-  -->
-
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-
-  <parent>
-    <groupId>org.optaweb.vehiclerouting</groupId>
-    <artifactId>plugins</artifactId>
-    <version>8.3.0-SNAPSHOT</version>
-  </parent>
-
-  <artifactId>websocket</artifactId>
-  <name>backend :: plugins :: routing</name>
-
-  <dependencies>
-    <dependency>
-      <groupId>org.optaweb.vehiclerouting</groupId>
-      <artifactId>domain</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.optaweb.vehiclerouting</groupId>
-      <artifactId>demo</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.optaweb.vehiclerouting</groupId>
-      <artifactId>error</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.optaweb.vehiclerouting</groupId>
-      <artifactId>region</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.optaweb.vehiclerouting</groupId>
-      <artifactId>route</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.springframework.boot</groupId>
-      <artifactId>spring-boot-starter-websocket</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.springframework.boot</groupId>
-      <artifactId>spring-boot-starter-test</artifactId>
-      <scope>test</scope>
-    </dependency>
-  </dependencies>
-</project>
