+++ optaweb-vehicle-routing-backend/pom.xml
@@ -89,10 +89,6 @@
       <groupId>org.optaweb.vehiclerouting</groupId>
       <artifactId>routing</artifactId>
     </dependency>
-    <dependency>
-      <groupId>org.optaweb.vehiclerouting</groupId>
-      <artifactId>websocket</artifactId>
-    </dependency>
     <!-- Production dependencies -->
     <dependency>
       <groupId>org.springframework.boot</groupId>
