+++ pom.xml
@@ -133,11 +133,6 @@
       </dependency>
       <dependency>
         <groupId>org.optaweb.vehiclerouting</groupId>
-        <artifactId>websocket</artifactId>
-        <version>${project.version}</version>
-      </dependency>
-      <dependency>
-        <groupId>org.optaweb.vehiclerouting</groupId>
         <artifactId>optaweb-vehicle-routing-backend</artifactId>
         <version>${project.version}</version>
       </dependency>
@@ -176,6 +171,23 @@
         <groupId>com.neovisionaries</groupId>
         <artifactId>nv-i18n</artifactId>
         <version>${version.com.neovisionaries}</version>
+      </dependency>
+      <!-- JSON testing -->
+      <dependency>
+        <groupId>org.springframework.boot</groupId>
+        <artifactId>spring-boot-test</artifactId>
+        <version>2.2.6.RELEASE</version>
+        <scope>test</scope>
+      </dependency>
+      <dependency>
+        <groupId>com.jayway.jsonpath</groupId>
+        <artifactId>json-path</artifactId>
+        <version>2.4.0</version>
+      </dependency>
+      <dependency>
+        <groupId>org.skyscreamer</groupId>
+        <artifactId>jsonassert</artifactId>
+        <version>1.5.0</version>
       </dependency>
     </dependencies>
   </dependencyManagement>
